refactor location models to own file
