+++ app/tests/test_views.py
@@ -12,7 +12,8 @@ from django.contrib.messages.storage.fallback import FallbackStorage
 from django.contrib.auth.models import User, Permission
 from django.contrib.contenttypes.models import ContentType
 from allauth.socialaccount.models import SocialAccount
-from mb.models.models import EntityClass, MasterReference, SourceAttribute, SourceEntity, SourceLocation, SourceMethod, SourceReference, SourceStatistic, TimePeriod, DietSet, FoodItem, DietSetItem, TaxonomicUnits, ChoiceValue, MasterEntity
+from mb.models.models import EntityClass, MasterReference, SourceAttribute, SourceEntity, SourceMethod, SourceReference, SourceStatistic, TimePeriod, DietSet, FoodItem, DietSetItem, TaxonomicUnits, ChoiceValue, MasterEntity
+from mb.models.location_models import SourceLocation
 from itis.models import Kingdom, TaxonUnitTypes
 from imports.checker import Check
 import imports.tools as tools
